refactor(dev-mode): extract seed into shippable DevModeSeed
